requirements.txt
pyspark==3.5.0
kafka-python>=2.0.2